ocall: add an example to show how to use OCALLs
The example first initializes a TEE context using the new settings API. It
then opens a session against the example OCALL TA via the new API to
configure sessions. Thereafter, it performs a function invocation into the
OCALL TA to request that the latter call it back.

The OCALL includes value and memref parameters that are passed in multiple
directions. Via these parameters the TA and CA exchange information the
same way they would during a normal CA-to-TA function invocation.

Signed-off-by: Hernan Gatta <hegatta@microsoft.com>